OperatorSkipTimed
